Add files via upload
adminLogin is where the administration logs on to there page. The adminLoginScript is the script I used to check the database and see if the information that was input was correct. The addEmployee allows an administration member to add an employee to the database and the deleteEmployee allows Administration to delete an employee.